pyproject.toml
[tool.poetry]
name = "todo-merger"
version = "0.1.0"
description = "To be filled"
# repository = ""
authors = ["Max Mehl <mail@deutschebahn.com>"]
readme = "README.md"
keywords = []
# license = ""
classifiers = []
packages = [{ include = "todo_merger" }]

[tool.poetry.scripts]
todo-merger = "todo_merger.__init__:main"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pygithub = "^2.3.0"
python-gitlab = "^4.7.0"
libsass = "^0.23.0"
toml = "^0.10.2"
platformdirs = "^4.2.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
pylint = "^3.2.5"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# FORMATTING settings
[tool.isort]
profile = "black"

[tool.black]
line-length = 100

# MYPY settings
[tool.mypy]
files = ["todo_merger/*.py"]
disable_error_code = ["import-untyped"]
